+++ mitmproxy/addons/view.py
@@ -497,7 +497,7 @@ class View(collections.abc.Sequence):
         Load flows into the view, without processing them with addons.
         """
         try:
-            with open(path, "rb") as f:
+            with io.open_flow_file(path) as f:
                 for i in io.FlowReader(f).stream():
                     # Do this to get a new ID, so we can load the same file N times and
                     # get new flows each time. It would be more efficient to just have a